pyproject.toml
[tool.poetry]
name = "ubiquitous-guide"
version = "0.1.0"
description = "Travel application to create and explore tourist spots around the world."
authors = ["Patrick Fourchalk <patrick4chalk@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.7"
psycopg = "^3.1.19"
black = "^24.8.0"
isort = "^5.13.2"
django-debug-toolbar = "^4.4.6"

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
line_length = 88
profile = "black"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
